# ⚡ Improve performance of `ln.connect()`, `lamin connect`, and `lamin load` for a notebook