This PR improves the CLI and documents profiling for the main PR below:

- https://github.com/laminlabs/lamindb/pull/1998

The initial profiling run below shows that `lamin load` for a notebook takes 4.7s for an instance (`laminlabs/lamindata`) that's in us-east-1 when called from a laptop in Munich.

In there, `ln.connect()` contributes almost all waiting time with **4.3s**.

<details><summary>Details</summary>
<p>

```
% kernprof -l lamin load https://lamin.ai/laminlabs/lamindata/transform/3qIZOmDUVBwc0000
→ connected lamindb: laminlabs/lamindata
→ updated uid: 3qIZOmDUVBwc0000 → 3qIZOmDUVBwc0001
→ notebook is here: demo.ipynb
Wrote profile results to lamin.lprof
Inspect results with:
python -m line_profiler -rmt "lamin.lprof"
% python -m line_profiler -rmt "lamin.lprof"
Timer unit: 1e-06 s

Total time: 4.70432 s
File: /Users/falexwolf/repos/laminhub/rest-hub/sub/lamindb/sub/lamin-cli/lamin_cli/_load.py
Function: load at line 19

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
    19                                           @profile                                                                                     
    20                                           def load(entity: str, uid: str = None, key: str = None, with_env: bool = False):             
    21         1          1.0      1.0      0.0      if entity.startswith("https://") and "lamin" in entity:                                  
    22         1          0.0      0.0      0.0          url = entity                                                                         
    23         1          3.0      3.0      0.0          instance_slug, entity, uid = decompose_url(url)                                      
    24                                               elif entity not in {"artifact", "transform"}:                                            
    25                                                   raise SystemExit("Entity has to be a laminhub URL or 'artifact' or 'transform'")     
    26                                               else:                                                                                    
    27                                                   instance_slug = None                                                                 
    28                                                                                                                                        
    29         1          0.0      0.0      0.0      if instance_slug is not None:                                                            
    30         1         18.0     18.0      0.0          auto_connect = ln_setup.settings.auto_connect                                        
    31                                                   # we don't want to auto-connect when importing lamindb                               
    32         1         80.0     80.0      0.0          ln_setup.settings.auto_connect = False                                               
    33                                                                                                                                        
    34         1       3677.0   3677.0      0.1          import lamindb as ln                                                                 
    35         1        415.0    415.0      0.0          from lamindb._finish import script_to_notebook                                       
    36                                                                                                                                        
    37         1         72.0     72.0      0.0          ln_setup.settings.auto_connect = auto_connect                                        
    38         1    4211566.0    4e+06     89.5          ln.connect(instance_slug)                                                            
    39                                               else:                                                                                    
    40                                                   import lamindb as ln                                                                 
    41                                                   from lamindb._finish import script_to_notebook                                       
    42                                                                                                                                        
    43                                               # below is to silence warnings about missing run inputs                                  
    44         1          2.0      2.0      0.0      ln.settings.track_run_inputs = False                                                     
    45                                                                                                                                        
    46         1          0.0      0.0      0.0      if entity == "transform":                                                                
    47         1          0.0      0.0      0.0          transform = (                                                                        
    48         1     133258.0 133258.0      2.8              ln.Transform.get(uid)                                                            
    49         1          0.0      0.0      0.0              if uid is not None                                                               
    50                                                       # if below, we take is_latest=True as the criterion, we might get draft notebooks
    51                                                       # hence, we use source_code__isnull=False and order by created_at instead        
    52                                                       else ln.Transform.filter(key=key, source_code__isnull=False)                     
    53                                                       .order_by("-created_at")                                                         
    54                                                       .first()                                                                         
    55                                                   )                                                                                    
    56         1          0.0      0.0      0.0          target_filename = transform.key                                                      
    57                                                   # if Path(target_filename).exists():                                                 
    58                                                   #     response = input(f"! {target_filename} exists: replace? (y/n)")                
    59                                                   #     if response != "y":                                                            
    60                                                   #         raise SystemExit("Aborted.")                                               
    61         1          2.0      2.0      0.0          if transform._source_code_artifact_id is not None:                                   
    62                                                       # backward compat                                                                
    63                                                       filepath_cache = transform._source_code_artifact.cache()                         
    64                                                       if not target_filename.endswith(transform._source_code_artifact.suffix):         
    65                                                           target_filename += transform._source_code_artifact.suffix                    
    66                                                       filepath_cache.rename(target_filename)                                           
    67         1          0.0      0.0      0.0          elif transform.source_code is not None:                                              
    68         1          1.0      1.0      0.0              if transform.key.endswith(".ipynb"):                                             
    69         1     355134.0 355134.0      7.5                  script_to_notebook(transform, target_filename, bump_revision=True)           
    70                                                       else:                                                                            
    71                                                           Path(target_filename).write_text(transform.source_code)                      
    72                                                   else:                                                                                
    73                                                       raise SystemExit("No source code available for this transform.")                 
    74         1         90.0     90.0      0.0          logger.important(f"{transform.type} is here: {target_filename}")                     
    75         1          0.0      0.0      0.0          if with_env:                                                                         
    76                                                       if (                                                                             
    77                                                           transform.latest_run is not None                                             
    78                                                           and transform.latest_run.environment is not None                             
    79                                                       ):                                                                               
    80                                                           filepath_env_cache = transform.latest_run.environment.cache()                
    81                                                           target_env_filename = (                                                      
    82                                                               ".".join(target_filename.split(".")[:-1]) + "__requirements.txt"         
    83                                                           )                                                                            
    84                                                           filepath_env_cache.rename(target_env_filename)                               
    85                                                           logger.important(target_env_filename)                                        
    86                                                       else:                                                                            
    87                                                           logger.warning("latest transform run with environment doesn't exist")        
    88         1          0.0      0.0      0.0          return target_filename                                                               
    89                                               elif entity == "artifact":                                                               
    90                                                   artifact = ln.Artifact.get(uid) if uid is not None else ln.Artifact.get(key=key)     
    91                                                   cache_path = artifact.cache()                                                        
    92                                                   logger.important(f"artifact is here: {cache_path}")                                  
    93                                                   return cache_path                                                                    


  4.70 seconds - /Users/falexwolf/repos/laminhub/rest-hub/sub/lamindb/sub/lamin-cli/lamin_cli/_load.py:19 - load
```

</p>
</details>