@falexwolf 2024-09-29
The roughly ~4sec have been what we've seen for a while.

https://laminlabs.slack.com/archives/C03P6D8U1PC/p1727270298402879?thread_ts=1727186286.045239&cid=C03P6D8U1PC
https://laminlabs.slack.com/archives/C03P6D8U1PC/p1727270024452489?thread_ts=1727186286.045239&cid=C03P6D8U1PC